realtime_chatapp/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-rest-knox==4.2.0
djangorestframework==3.15.1
pycparser==2.22
sqlparse==0.5.0
tzdata==2024.1
